# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness App API
  version: 1.0.0
  description: |
    REST API za fitness aplikaciju (registracija, login, profil, treninzi, napredak).
servers:
  - url: http://localhost:8080
    description: Lokalni backend server

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Server radi
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /api/register:
    post:
      summary: Registracija novog korisnika
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Uspešna registracija, vraća korisnika i JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Neispravan zahtev
        '409':
          description: Email već postoji

  /api/login:
    post:
      summary: Prijava korisnika
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Uspešna prijava, vraća korisnika i JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Pogrešan email ili lozinka

  /api/logout:
    post:
      summary: Odjava korisnika
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Uspešna odjava

  /api/profile:
    get:
      summary: Dohvatanje profila ulogovanog korisnika
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Podaci o korisniku
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Neautorizovano

  /api/workouts:
    get:
      summary: Lista treninga za korisnika
      tags: [Workouts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista treninga
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workout'

  /api/workouts/create:
    post:
      summary: Kreiranje novog treninga
      tags: [Workouts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRequest'
      responses:
        '200':
          description: Kreiran trening
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          description: Neispravan zahtev
        '401':
          description: Neautorizovano

  /api/workouts/update:
    put:
      summary: Ažuriranje treninga
      tags: [Workouts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutRequest'
      responses:
        '200':
          description: Ažuriran trening
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          description: Neispravan zahtev
        '401':
          description: Neautorizovano

  /api/workouts/delete:
    delete:
      summary: Brisanje treninga
      tags: [Workouts]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
          description: ID treninga
      responses:
        '200':
          description: Obrisan trening

  /api/progress:
    get:
      summary: Lista zapisa napretka
      tags: [Progress]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista zapisa napretka
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressEntry'

  /api/progress/create:
    post:
      summary: Kreiranje zapisa napretka
      tags: [Progress]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressRequest'
      responses:
        '200':
          description: Kreiran zapis napretka
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressEntry'
        '400':
          description: Neispravan zahtev
        '401':
          description: Neautorizovano

  /api/progress/update:
    put:
      summary: Ažuriranje zapisa napretka
      tags: [Progress]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressRequest'
      responses:
        '200':
          description: Ažuriran zapis napretka
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressEntry'
        '400':
          description: Neispravan zahtev
        '401':
          description: Neautorizovano

  /api/progress/delete:
    delete:
      summary: Brisanje zapisa napretka
      tags: [Progress]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
          description: ID zapisa napretka
      responses:
        '200':
          description: Obrisan zapis napretka

  /api/food/search:
    post:
      summary: Pretraga hrane po barcodu
      tags: [Food]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodSearchRequest'
      responses:
        '200':
          description: Pronađena hrana
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '404':
          description: Proizvod nije pronađen
        '401':
          description: Neautorizovano

  /api/meal-plan:
    get:
      summary: Generisanje plana ishrane
      tags: [Food]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Plan ishrane
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlan'
        '401':
          description: Neautorizovano

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        goal:
          type: string
          enum: [lose_weight, hypertrophy]
        role:
          type: string
          example: user
        height:
          type: number
          format: float
          nullable: true
        weight:
          type: number
          format: float
          nullable: true

    RegisterRequest:
      type: object
      required: [name, email, password, goal]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        goal:
          type: string
          enum: [lose_weight, hypertrophy]
        height:
          type: number
          format: float
          nullable: true
        weight:
          type: number
          format: float
          nullable: true

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string

    Workout:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: integer
          description: Trajanje u minutama
        calories_burned:
          type: number
          format: float
          description: Kalorije sagorene
        workout_date:
          type: string
          format: date
          description: Datum treninga
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProgressEntry:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        weight:
          type: number
          format: float
          description: Težina u kg
        body_fat:
          type: number
          format: float
          nullable: true
          description: Procenat masti u telu
        muscle_mass:
          type: number
          format: float
          nullable: true
          description: Mišićna masa u kg
        notes:
          type: string
          nullable: true
          description: Napomene
        progress_date:
          type: string
          format: date
          description: Datum zapisa napretka
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    FoodSearchRequest:
      type: object
      required: [barcode]
      properties:
        barcode:
          type: string
          description: Barcode proizvoda

    Food:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: string
        calories:
          type: number
          format: float
        protein:
          type: number
          format: float
        carbs:
          type: number
          format: float
        fat:
          type: number
          format: float

    MealPlan:
      type: object
      properties:
        userId:
          type: integer
        goal:
          type: string
          enum: [lose_weight, hypertrophy]
        foods:
          type: array
          items:
            $ref: '#/components/schemas/Food'
        totalCalories:
          type: number
          format: float
        totalProtein:
          type: number
          format: float
        totalCarbs:
          type: number
          format: float
        totalFat:
          type: number
          format: float

    WorkoutRequest:
      type: object
      required: [name, duration, calories_burned, workout_date]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        duration:
          type: integer
          minimum: 1
          description: Trajanje u minutama
        calories_burned:
          type: number
          format: float
          minimum: 0
          description: Kalorije sagorene
        workout_date:
          type: string
          format: date
          description: Datum treninga (YYYY-MM-DD)

    ProgressRequest:
      type: object
      required: [weight, progress_date]
      properties:
        weight:
          type: number
          format: float
          minimum: 0
          description: Težina u kg
        body_fat:
          type: number
          format: float
          minimum: 0
          maximum: 100
          nullable: true
          description: Procenat masti u telu
        muscle_mass:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Mišićna masa u kg
        notes:
          type: string
          nullable: true
          description: Napomene
        progress_date:
          type: string
          format: date
          description: Datum zapisa napretka (YYYY-MM-DD)